requirements.txt
discord.py
pandas
requests
playwright
openpyxl
tabulate
matplotlib
seaborn
jinja2
python-dotenv
